FE: Convert the concrete http client to a singleton
So here's the deal. I want to redirect unauthorized users to '/login' when I receive a 401 response from the backend.

Since this is a react app, I can't use the navigation hooks outside of react components. In my use case here, I want to use the router in some .ts file that's not related to react.

Now since this is not possible, I created an array of listeners in the http client. When I receive an error in the API response, I will notify the subscribers -- including a specific one for routing purposes. When this component gets notified, it checks for the error status and redirects if it's a 401 error.

So why use singleton? Because I want a single array of error listeners for the whole application.